move spark utils, clean up
