+++ caikit/interfaces/ts/data_model/backends/util.py
@@ -15,7 +15,7 @@
 
 # Standard
 from datetime import datetime
-from typing import Any, Iterable, List, Union
+from typing import Any, Iterable, Union
 
 # Third Party
 import numpy as np
@@ -23,23 +23,6 @@ import pandas as pd
 
 # Local
 from ..toolkit.optional_dependencies import HAVE_PYSPARK, pyspark
-
-
-def mock_pd_groupby(a_df_like, by: List[str], return_pandas_api=False):
-    """Roughly mocks the behavior of pandas groupBy but on a spark dataframe."""
-
-    distinct_keys = a_df_like.select(by).distinct().collect()
-    for dkey in distinct_keys:
-        adict = dkey.asDict()
-        filter_statement = ""
-        for k, v in adict.items():
-            filter_statement += f" {k} == '{v}' and"
-        if filter_statement.endswith("and"):
-            filter_statement = filter_statement[0:-3]
-        sub_df = a_df_like.filter(filter_statement)
-        value = tuple(adict.values())
-        value = value[0] if len(value) == 1 else value
-        yield value, sub_df.pandas_api() if return_pandas_api else sub_df
 
 
 def timezoneoffset(adatetime: datetime) -> int:
